Refactor: move LeaderState to leader_state.rs
